requirements.txt
numpy
scipy
pandas
rich
pytest
optuna
pybaseball
MLB-StatsAPI
